requirements.txt
python-pptx
transformer